Move updateInsidePolyfill_works to SharedDriverConformanceTest
