+++ bosk-testing/src/main/java/works/bosk/drivers/state/UpgradeableEntity.java
@@ -1,0 +1,27 @@
+package works.bosk.drivers.state;
+
+import java.util.Optional;
+import works.bosk.Catalog;
+import works.bosk.Entity;
+import works.bosk.Identifier;
+import works.bosk.Listing;
+import works.bosk.SideTable;
+import works.bosk.TaggedUnion;
+import works.bosk.annotations.Polyfill;
+
+/**
+ * A version of {@link TestEntity} where the {@link Optional} {@link TestEntity#values()}
+ * field has a polyfill.
+ */
+public record UpgradeableEntity(
+	Identifier id,
+	String string,
+	Catalog<TestEntity> catalog,
+	Listing<TestEntity> listing,
+	SideTable<TestEntity, TestEntity> sideTable,
+	TaggedUnion<TestEntity.Variant> variant,
+	TestValues values
+) implements Entity {
+	@Polyfill("values")
+	public static final TestValues DEFAULT_VALUES = TestValues.blank();
+}
